backend/app/pyproject.toml
[tool.poetry]
name = "shinan-backend"
version = "0.1.0"
description = "Backend API for financial research"
authors = ["Kushal Chattopadhyay <kushalfremont@gmail.com>"]
readme = "README.md"
package-mode = false  # disables packaging since it's an app

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115"
uvicorn = "^0.34"
pydantic = "^2.0"
rich = "^13.0"
openai = "^1.0"
aiohttp = "^3.8"
httpx = ">=0.27,<1.0.0"
python-multipart = ">=0.0.9,<1.0.0"
openai-agents = "^0.0.19"
PyPDF2 = "^3.0.0"
PyMuPDF = "^1.23.0"
pytesseract = "^0.3.10"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"